Reorganize reporting submodules
