+++ ixmp/reporting/exceptions.py
@@ -1,0 +1,61 @@
+
+class ComputationError(Exception):
+    """Wrapper to print intelligible exception information for Reporter.get().
+
+    In order to aid in debugging, this helper:
+    - Omits the parts of the stack trace that are internal to dask, and
+    - Gives the key in the Reporter.graph and the computation task that
+      caused the exception.
+    """
+    def __str__(self):
+        from traceback import (
+            TracebackException,
+            format_exception_only,
+            format_list,
+        )
+
+        # Move the cause to a non-private attribute
+        self.cause = self.__cause__
+
+        # Suppress automatic printing of the cause
+        self.__cause__ = None
+
+        info = None  # Information about the call that triggered the exception
+        frames = []  # Frames for an abbreviated stacktrace
+        dask_internal = True  # Flag if the frame is internal to dask
+
+        # Iterate over frames from the base of the stack
+        tb = TracebackException.from_exception(self.cause, capture_locals=True)
+        for frame in tb.stack:
+            if frame.name == 'execute_task':
+                # Current frame is the dask internal call to execute a task
+
+                # Retrieve information about the key/task that triggered the
+                # exception. These are not the raw values of variables, but
+                # their string repr().
+                info = {name: frame.locals[name] for name in ('key', 'task')}
+
+                # Remaining frames are related to the exception
+                dask_internal = False
+
+            if not dask_internal:
+                # Don't display the locals when printing the traceback
+                frame.locals = None
+
+                # Store the frame for printing the traceback
+                frames.append(frame)
+
+        # Assemble the exception printout
+
+        # Reporter information for debugging
+        lines = [
+            'when computing {key}, using\n\n{task}\n\n'.format(**info),
+            'Use Reporter.describe(...) to trace the computation.\n\n',
+            'Computation traceback:\n',
+        ]
+        # Traceback; omitting a few dask internal calls below execute_task
+        lines.extend(format_list(frames[3:]))
+        # Type and message of the original exception
+        lines.extend(format_exception_only(self.cause.__class__, self.cause))
+
+        return ''.join(lines)
