+++ tests/data/report-describe.txt
@@ -4,32 +4,32 @@
   - data_for_quantity('par', 'a', 'value', ...)
   - 'scenario':
     - <ixmp.core.Scenario object at {id}>
-  - 'filters':
-    - {{}}
+  - 'config':
+    - {{'filters': {{}}}}
 
 'b:':
 - sum(dimensions=['j'], weights=None, ...)
 - 'b:j':
   - data_for_quantity('par', 'b', 'value', ...)
   - 'scenario' (above)
-  - 'filters' (above)
+  - 'config' (above)
 
 'cost-margin:':
 - data_for_quantity('equ', 'cost', 'mrg', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'cost:':
 - data_for_quantity('equ', 'cost', 'lvl', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'd:':
 - sum(dimensions=['i', 'j'], weights=None, ...)
 - 'd:i-j':
   - data_for_quantity('par', 'd', 'value', ...)
   - 'scenario' (above)
-  - 'filters' (above)
+  - 'config' (above)
 
 'd:i':
 - sum(dimensions=['j'], weights=None, ...)
@@ -42,19 +42,19 @@
 'demand-margin:j':
 - data_for_quantity('equ', 'demand', 'mrg', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'demand:':
 - sum(dimensions=['j'], weights=None, ...)
 - 'demand:j':
   - data_for_quantity('equ', 'demand', 'lvl', ...)
   - 'scenario' (above)
-  - 'filters' (above)
+  - 'config' (above)
 
 'f:':
 - data_for_quantity('par', 'f', 'value', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'i':
 - ['seattle', 'san-diego']
@@ -65,21 +65,21 @@
 'supply-margin:i':
 - data_for_quantity('equ', 'supply', 'mrg', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'supply:':
 - sum(dimensions=['i'], weights=None, ...)
 - 'supply:i':
   - data_for_quantity('equ', 'supply', 'lvl', ...)
   - 'scenario' (above)
-  - 'filters' (above)
+  - 'config' (above)
 
 'x:':
 - sum(dimensions=['i', 'j'], weights=None, ...)
 - 'x:i-j':
   - data_for_quantity('var', 'x', 'lvl', ...)
   - 'scenario' (above)
-  - 'filters' (above)
+  - 'config' (above)
 
 'x:i':
 - sum(dimensions=['j'], weights=None, ...)
@@ -92,7 +92,7 @@
 'z:':
 - data_for_quantity('var', 'z', 'lvl', ...)
 - 'scenario' (above)
-- 'filters' (above)
+- 'config' (above)
 
 'all':
 - list of:
